+++ test/test_dash_validate.py
@@ -13,16 +13,15 @@ dash_sparql_files = []
 dash_triple_rules_files = []
 dash_sparql_rules_files = []
 dash_target_files = []
+dash_fn_files = []
 
 # There are some tests we know will fail, but we don't want to stop deployment
 # if we hit them. List them here:
-ALLOWABLE_NOT_IMPLEMENTED = [
-    "/rules/triple/rectangle.test.ttl",
-    "/rules/triple/person2schema.test.ttl",
-    "/rules/triple/functions-permutations.test.ttl",
-]
+ALLOWABLE_NOT_IMPLEMENTED = []
 
-ALLOWABLE_FAILURES = []
+# This one relies on owl_imports to be turned on
+# and also needs this file, but its missing: http://datashapes.org/shasf/tests/rules/triple/person
+ALLOWABLE_FAILURES = ["/rules/triple/person2schema.test.ttl"]
 
 for x in walk(path.join(dash_files_dir, 'core')):
     for y in glob.glob(path.join(x[0], '*.test.ttl')):
@@ -145,6 +144,47 @@ for x in walk(path.join(dash_files_dir, 'target')):
         dash_target_files.append((y, None))
 @pytest.mark.parametrize('target_file, shacl_file', dash_target_files)
 def test_dash_validate_target(target_file, shacl_file):
+    test_name = shacl_file or target_file
+    try:
+        val, _, v_text = pyshacl.validate(
+            target_file, shacl_graph=shacl_file, advanced=True, inference='rdfs', check_dash_result=True, debug=True, meta_shacl=False)
+    except NotImplementedError as ne:
+        for ani in ALLOWABLE_NOT_IMPLEMENTED:
+            if test_name.endswith(ani):
+                v_text = "Skipping not implemented feature in test: {}".format(test_name)
+                print(v_text)
+                val = True
+                break
+        else:
+            print(ne)
+            val = False
+            v_text = ""
+    except ReportableRuntimeError as e:
+        import traceback
+        print(e)
+        traceback.print_tb(e.__traceback__)
+        val = False
+        v_text = ""
+    try:
+        assert val
+    except AssertionError as ae:
+        for af in ALLOWABLE_FAILURES:
+            if test_name.endswith(af):
+                v_text = "Allowing failure in test: {}".format(test_name)
+                print(v_text)
+                break
+        else:
+            raise ae
+
+    print(v_text)
+    return True
+
+# Get all SHACLFunction tests
+for x in walk(path.join(dash_files_dir, 'function')):
+    for y in glob.glob(path.join(x[0], '*.test.ttl')):
+        dash_fn_files.append((y, None))
+@pytest.mark.parametrize('target_file, shacl_file', dash_fn_files)
+def test_dash_validate_functions(target_file, shacl_file):
     test_name = shacl_file or target_file
     try:
         val, _, v_text = pyshacl.validate(